Refactor rule changed subscriber
