+++ features/readwrite-splitting/core/src/main/resources/META-INF/services/org.apache.shardingsphere.mode.subsciber.RuleChangedSubscriber
